refactor: 핸들러 리팩토링 공사 완
